build the bs flags only once

+++ lib/motion/project/template/ios/builder.rb
@@ -92,8 +92,10 @@ PLIST
 
       # Copy the provisioning profile.
       bundle_provision = File.join(bundle_path, "embedded.mobileprovision")
-      App.info 'Create', bundle_provision
-      FileUtils.cp config.provisioning_profile, bundle_provision
+      if !File.exist?(bundle_provision) or File.mtime(config.provisioning_profile) > File.mtime(bundle_provision)
+        App.info 'Create', bundle_provision
+        FileUtils.cp config.provisioning_profile, bundle_provision
+      end
 
       # Codesign.
       codesign_cmd = "CODESIGN_ALLOCATE=\"#{File.join(config.platform_dir(platform), 'Developer/usr/bin/codesign_allocate')}\" /usr/bin/codesign"
